requirements.txt
aiohttp>=3.9,<4
rapidfuzz>=3.6,<4
python-dotenv>=1.0,<2
pydantic>=2.5,<3
pytest>=8.0,<9
pytest-asyncio>=0.23,<1
scipy>=1.12,<2
pulp>=2.8,<3
cvxpy>=1.4,<2
numpy>=1.26,<3
